font_families_google {sysfonts}	R Documentation

List Font Families Available in Google Fonts

Description

The two versions of this function are equivalent, but the "underscore" naming is preferred.

This function lists family names of the fonts that are currently available in Google Fonts. When running this function for the first time, it may take a few seconds to fetch the font information database. This function requires the jsonlite and curl packages.

Usage

font_families_google(db_cache = TRUE, handle = curl::new_handle())

font.families.google()

Arguments

db_cache
whether to obtain font metadata from a cache site. Using cache is typically faster, but not as update-to-date as using the official API. If db_cache is set to FALSE, then metadata are retrieved from the Google Fonts API.

handle
a curl handle object passed to curl::curl_download() and curl::curl_fetch_memory().

Value

A character vector of available font family names in Google Fonts.

Author(s)

Yixuan Qiu <https://statr.me/>

See Also

font_add_google()

Examples

## Not run: 
font_families_google()

## End(Not run)

[Package sysfonts version 0.8.3 Index]