font_files {sysfonts}	R Documentation

List Font Files Available in the Search Paths

Description

The two versions of this function are equivalent, but the "underscore" naming is preferred.

This function lists font files in the search path that can be loaded by font_add(). Currently supported formats include TrueType fonts(*.ttf, *.ttc) and OpenType fonts(*.otf).

Usage

font_files()

font.files()

Value

A data frame containing the following information of the font files:

path
The directory that the font file is located in.

file
File name of the font.

family
Family name.

face
Font face.

version
Version of the font.

ps_name
PostScript font name.

Author(s)

Yixuan Qiu <https://statr.me/>

See Also

font_paths(), font_add()

Examples

## Not run: 
font_files()

## End(Not run)

[Package sysfonts version 0.8.3 Index]